// locations-frontend/src/api/endpoints/locations/locations.zod.ts
export const locationsControllerUpdateBodyNameMin = 2;
export const locationsControllerUpdateBodyNameMax = 60;

export const locationsControllerUpdateBodyCoordinatesLonMin = -180;
export const locationsControllerUpdateBodyCoordinatesLonMax = 180;

export const locationsControllerUpdateBodyCoordinatesLatMin = -90;
export const locationsControllerUpdateBodyCoordinatesLatMax = 90;

export const locationsControllerUpdateBodyAddressMax = 120;

export const locationsControllerUpdateBodyNotesMax = 500;



export const LocationsControllerUpdateBody = zod.object({
  "name": zod.string().min(locationsControllerUpdateBodyNameMin).max(locationsControllerUpdateBodyNameMax).optional(),
  "category": zod.enum(['office', 'store', 'landmark']).optional(),
  "coordinates": zod.object({
  "lon": zod.number().min(locationsControllerUpdateBodyCoordinatesLonMin).max(locationsControllerUpdateBodyCoordinatesLonMax),
  "lat": zod.number().min(locationsControllerUpdateBodyCoordinatesLatMin).max(locationsControllerUpdateBodyCoordinatesLatMax)
}).optional(),
  "address": zod.string().max(locationsControllerUpdateBodyAddressMax).optional(),
  "notes": zod.string().max(locationsControllerUpdateBodyNotesMax).optional()
})