// locations-frontend/src/api/models/locationDto.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Locations
 * The locations API description
 * OpenAPI spec version: 1.0
 */
import type { LocationDtoCategory } from './locationDtoCategory';
import type { LocationDtoCoordinates } from './locationDtoCoordinates';

export interface LocationDto {
  id: string;
  /**
     * @minLength 2
     * @maxLength 60
     */
  name: string;
  category: LocationDtoCategory;
  coordinates: LocationDtoCoordinates;
  /** @maxLength 120 */
  address?: string;
  /** @maxLength 500 */
  notes?: string;
  /** @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$ */
  createdAt: string;
  /** @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$ */
  updatedAt: string;
}